Added checking of creation name to ensure no illegal chars

+++ src/application/CreationPreviewController.java
@@ -1,10 +1,7 @@
 package application;
 
 import javafx.fxml.FXML;
-import javafx.scene.control.Alert;
-import javafx.scene.control.Button;
-import javafx.scene.control.ButtonType;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.media.Media;
 import javafx.scene.media.MediaPlayer;
 import javafx.scene.media.MediaView;
@@ -23,6 +20,8 @@ public class CreationPreviewController {
     private Button confirmButton;
     @FXML
     private TextField nameField;
+    @FXML
+    private Label nameErrorLabel;
 
     @FXML
     private void initialize(){
@@ -37,35 +36,38 @@ public class CreationPreviewController {
     @FXML
     private void confirmButtonClicked() throws InterruptedException, IOException {
         String creationName = nameField.getText();
-        ProcessBuilder pb = new ProcessBuilder();
-        pb.command("/bin/bash", "-c", "[ -e ./creations/\"" + creationName + "\" ]");
-        Process checkExist = pb.start();
-        int exit = checkExist.waitFor();
+        if (nameIsValid(creationName)) {
+            ProcessBuilder pb = new ProcessBuilder();
+            pb.command("/bin/bash", "-c", "[ -e ./creations/\"" + creationName + "\" ]");
+            Process checkExist = pb.start();
+            int exit = checkExist.waitFor();
 
-        if (exit == 1) {
-            pb.command("/bin/bash", "-c", "mv ./.temp ./creations/" + "\"" + creationName + "\"");
-            Process moveTemp = pb.start();
-            Stage thisStage = (Stage)nameField.getScene().getWindow();
-            thisStage.close();
-        } else {
-            Alert alert = new Alert(Alert.AlertType.NONE, "Creation: '" + creationName + "' already exists. Do you wish to overwrite it?", ButtonType.YES, ButtonType.NO);
-            alert.setTitle("Creation already exists");
-            alert.setHeight(150);
-            alert.showAndWait();
-
-            if (alert.getResult() == ButtonType.YES) {
-                pb.command("/bin/bash", "-c", "rm -fr ./creations/\"" + creationName + "\"");
-                Process removeExisting = pb.start();
-                removeExisting.waitFor();
-
-                pb.command("/bin/bash", "-c", "mv -f ./.temp ./creations/\"" + creationName + "\"");
-                pb.start();
-
+            if (exit == 1) {
+                pb.command("/bin/bash", "-c", "mv ./.temp ./creations/" + "\"" + creationName + "\"");
+                Process moveTemp = pb.start();
                 Stage thisStage = (Stage)nameField.getScene().getWindow();
                 thisStage.close();
-            }
-        }
+            } else {
+                Alert alert = new Alert(Alert.AlertType.NONE, "Creation: '" + creationName + "' already exists. Do you wish to overwrite it?", ButtonType.YES, ButtonType.NO);
+                alert.setTitle("Creation already exists");
+                alert.setHeight(150);
+                alert.showAndWait();
 
+                if (alert.getResult() == ButtonType.YES) {
+                    pb.command("/bin/bash", "-c", "rm -fr ./creations/\"" + creationName + "\"");
+                    Process removeExisting = pb.start();
+                    removeExisting.waitFor();
+
+                    pb.command("/bin/bash", "-c", "mv -f ./.temp ./creations/\"" + creationName + "\"");
+                    pb.start();
+
+                    Stage thisStage = (Stage)nameField.getScene().getWindow();
+                    thisStage.close();
+                }
+            }
+        } else {
+            nameErrorLabel.setVisible(true);
+        }
     }
 
     @FXML
@@ -85,6 +87,10 @@ public class CreationPreviewController {
             player.play();
             playPauseButton.setText("Pause");
         }
+    }
+
+    private boolean nameIsValid(String creationName) {
+        return creationName.matches("[A-Za-z0-9_\\-][A-Za-z0-9_\\- ]*");
     }
 
 }
